commit: checkpoint 6/4/2025 4:09 AM

+++ Assets/com/ethnicthv/chemlab/engine/api/atom/Atom.cs
@@ -21,7 +21,7 @@ namespace com.ethnicthv.chemlab.engine.api.atom
             return ElementProperty.GetElementProperty(_element);
         }
         
-        public int GetMaxConnectivity()
+        public double GetMaxConnectivity()
         {
             return GetProperty().Valences.Max() ;
         }
